Install/requirements/local.txt
-r base.txt

aiohttp
aiohttp-retry
aiosignal
asgiref
async-timeout
attrs
certifi
charset-normalizer
Django
django-ckeditor==6.3.2
django-js-asset
django-modeltranslation
django-user-agents
frozenlist
idna
multidict
pillow
PyJWT
requests
sqlparse
twilio
typing_extensions
ua-parser
urllib3
user-agents
whitenoise
yarl
psycopg2-binary
mysqlclient
